feito ajuste de valores no base stats

+++ app/src/main/java/com/bruno/pokedex/domain/model/PokemonDetail.kt
@@ -2,11 +2,11 @@ package com.bruno.pokedex.domain.model
 
 data class PokemonDetail(
     val baseExperience: Int,
-    val height: Int,
+    val height: Float,
     val id: Int,
     val name: String,
     val sprites: Sprites,
     val stats: List<StatsDetail>,
     val types: List<TypeDetail>,
-    val weight: Int
+    val weight: Float
 )
